# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "我的日程本API",
        "title": "library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
...
        "/board/searchdate": {
            "post": {
                "description": "输入年月日",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "date"
                ],
                "summary": "查日期",
                "parameters": [
                    {
                        "type": "string",
                        "description": "年",
                        "name": "year",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "月",
                        "name": "month",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日",
                        "name": "day",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SetDate"
                        }
                    }
                }
            }
        },
...
        "/preview/{year}/{month}/{day}": {
            "get": {
                "description": "查看当天的图片与做完的任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "door"
                ],
                "summary": "打开任意门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "年",
                        "name": "year",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "月",
                        "name": "month",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日",
                        "name": "day",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SchAndPh"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.SchAndPh": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "photo": {
                    "type": "object",
                    "properties": {
                        "day": {
                            "type": "integer"
                        },
                        "insertTime": {
                            "type": "string"
                        },
                        "month": {
                            "type": "integer"
                        },
                        "number": {
                            "type": "integer"
                        },
                        "photoUrl": {
                            "type": "string"
                        },
                        "userId": {
                            "type": "string"
                        },
                        "year": {
                            "type": "integer"
                        }
                    }
                },
                "schedule": {
                    "description": "我感觉这里的数据类型是要改的//明早再说好了",
                    "type": "object",
                    "properties": {
                        "content": {
                            "type": "string"
                        },
                        "day": {
                            "type": "integer"
                        },
                        "done": {
                            "type": "boolean"
                        },
                        "month": {
                            "type": "integer"
                        },
                        "userId": {
                            "type": "string"
                        },
                        "year": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "model.SetDate": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "day": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    }
}